garbage_collector/requirements.txt
imutils==0.5.3
numpy==1.16.6
opencv-python==4.2.0.32
pyserial==3.4
